refactor: PhotographerPackage 파일 분리
